requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
harvest-api==0.4.0
idna==3.4
requests==2.28.1
tabulate==0.9.0
urllib3==1.26.13
